Avoid filtering preferences by --reinstall
